# api-contract/openapi.v1.yaml
openapi: 3.0.3
info:
  title: AI Mentor API
  version: 2.0.0
servers:
  - url: /api/v1
paths:
  /health:
    get:
      summary: Service health and capabilities
      responses:
        "200":
          description: Healthy response
  /tools:
    get:
      summary: Available language toolchains
      responses:
        "200":
          description: Tool availability
  /analyze:
    post:
      summary: Analyze code and return issues and execution output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                language:
                  type: string
                  default: python
                difficulty:
                  type: string
                  enum: [beginner, intermediate, advanced]
                  default: beginner
      responses:
        "200":
          description: Successful analysis
        "400":
          description: Invalid request (missing code, too large)
        "422":
          description: Required toolchain unavailable
  /csrf-token:
    get:
      summary: Obtain a CSRF token for mutating requests
      responses:
        "200":
          description: CSRF token returned
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User created
        "400":
          description: Validation error
        "409":
          description: Email already taken
  /auth/login:
    post:
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful, tokens returned
        "401":
          description: Invalid credentials
  /auth/github:
    get:
      summary: Initiate GitHub OAuth login
      parameters:
        - name: redirect
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to GitHub authorization page
  /auth/github/callback:
    get:
      summary: GitHub OAuth callback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect with session cookie set
        "400":
          description: OAuth error
  /auth/me:
    get:
      summary: Get current authenticated user
      responses:
        "200":
          description: User object
        "401":
          description: Not authenticated
  /auth/logout:
    post:
      summary: Log out (clear refresh cookie)
      responses:
        "200":
          description: Logged out
  /history:
    get:
      summary: Get analyze history for current user
      responses:
        "200":
          description: List of history entries
        "401":
          description: Not authenticated
    delete:
      summary: Clear analyze history for current user
      responses:
        "200":
          description: History cleared
        "401":
          description: Not authenticated
  /metrics:
    get:
      summary: Prometheus-format security metrics (API-key gated in production)
      responses:
        "200":
          description: Metrics in Prometheus text format
  /debug/sandbox-status:
    get:
      summary: Sandbox runtime diagnostics (disabled in production)
      responses:
        "200":
          description: Sandbox diagnostic info
  /debug/gemini-status:
    get:
      summary: Gemini integration diagnostics (disabled in production)
      responses:
        "200":
          description: Diagnostic result